+++ previous-versions/accounting/RELEASES.md
@@ -719,3 +719,13 @@ Based on:
 - [typescript v0.41.0] previous-versions/accounting
 ### Releases
 - [NPM v0.41.0] https://www.npmjs.com/package/@codat/accounting/v/0.41.0 - previous-versions/accounting
+
+## 2023-12-15 12:54:46
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.126.3 (2.214.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.42.0] previous-versions/accounting
+### Releases
+- [NPM v0.42.0] https://www.npmjs.com/package/@codat/accounting/v/0.42.0 - previous-versions/accounting